requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
xlsxwriter>=3.1.0
numpy>=1.24.0
python-dateutil>=2.8.0
python-pptx>=0.6.21
fpdf2>=2.7.0
